requirements.txt
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.7
django-oauth-toolkit==1.0.0
django-storages==1.6.5
djangorestframework==3.6.3
djangorestframework-oauth==1.1.0
gunicorn==19.7.1
idna==2.6
oauth2==1.9.0.post1
oauthlib==2.0.6
piexif==1.0.13
Pillow==5.0.0
pytz==2017.3
requests==2.18.4
urllib3==1.22
whitenoise==3.3.1
